Remove enum and neaten
