+++ src/dotnet-roslyn/Commands/VSBranchInfoCommand.cs
@@ -14,9 +14,11 @@ internal class VSBranchInfoCommand
 {
     private static readonly VSBranchInfoCommandDefaultHandler s_vsBranchInfoCommandHandler = new();
 
-    internal static readonly Option<string> BranchOption = new Option<string>(new[] { "--branch", "-b" }, () => "main", "Which VS branch to show information for (eg main, rel/d17.1)");
-    internal static readonly Option<Product> ProductOption = new Option<Product>(new[] { "--product", "-p" }, () => Product.Roslyn, "Which product to get info for (roslyn or razor)");
-    internal static readonly Option ShowArtifacts = new Option(new[] { "--show-artifacts", "-a" }, "Whether to show artifact download links for the packages product by the build (if available)");
+    private static readonly string[] s_allProductNames = VSBranchInfo.AllProducts.Select(p => p.Name).Concat(new[] { "all" }).ToArray();
+
+    internal static readonly Option<string> BranchOption = new(new[] { "--branch", "-b" }, () => "main", "Which VS branch to show information for (eg main, rel/d17.1)");
+    internal static readonly Option<string> ProductOption = new Option<string>(new[] { "--product", "-p" }, () => "roslyn", "Which product to get info for (roslyn or razor)").FromAmong(s_allProductNames);
+    internal static readonly Option ShowArtifacts = new(new[] { "--show-artifacts", "-a" }, "Whether to show artifact download links for the packages product by the build (if available)");
 
     public static Symbol GetCommand()
     {
@@ -38,7 +40,7 @@ internal class VSBranchInfoCommand
             var logger = context.SetupLogging();
 
             var branch = context.ParseResult.GetValueForOption(BranchOption)!;
-            var product = context.ParseResult.GetValueForOption(ProductOption);
+            var product = context.ParseResult.GetValueForOption(ProductOption)!;
             var showArtifacts = context.ParseResult.WasOptionUsed(ShowArtifacts.Aliases.ToArray());
 
             return VSBranchInfo.GetInfoAsync(branch, product, showArtifacts, logger);
