--- src/dotnet-roslyn/VS/Product.cs
@@ -1,12 +1,0 @@
-// Licensed to the.NET Foundation under one or more agreements.
-// The.NET Foundation licenses this file to you under the MIT license.
-// See the License.txt file in the project root for more information.
-
-namespace Microsoft.Roslyn.VS;
-
-internal enum Product
-{
-    Roslyn,
-    Razor,
-    All
-}
